gnrc_ppp: lcp: Changed uint8_t to int in some functions with negative value

+++ sys/include/net/gnrc/ppp/lcp.h
@@ -50,10 +50,10 @@ void lcp_scj(ppp_cp_t *lcp, cp_pkt_t *pkt);
 void lcp_ser(ppp_cp_t *lcp, cp_pkt_t *pkt);
 
 /*Event generators*/
-uint8_t lcp_handle_conf(ppp_cp_t *lcp, cp_pkt_t *pkt);
-uint8_t lcp_handle_code(cp_pkt_t *pkt);
+int lcp_handle_conf(ppp_cp_t *lcp, cp_pkt_t *pkt);
+int lcp_handle_code(cp_pkt_t *pkt);
 
-uint8_t lcp_handle_pkt(ppp_cp_t *lcp, cp_pkt_t *pkt);
+int lcp_handle_pkt(ppp_cp_t *lcp, cp_pkt_t *pkt);
 
 #ifdef __cplusplus
 }
